Split ProfileService into ProfileService interface and impl
