+++ api/src/main/java/org/pmiops/workbench/profile/ProfileService.java
@@ -1,138 +1,11 @@
 package org.pmiops.workbench.profile;
 
-import java.util.Optional;
-import org.pmiops.workbench.billing.FreeTierBillingService;
-import org.pmiops.workbench.db.dao.InstitutionDao;
-import org.pmiops.workbench.db.dao.UserDao;
-import org.pmiops.workbench.db.dao.UserService;
-import org.pmiops.workbench.db.dao.UserTermsOfServiceDao;
-import org.pmiops.workbench.db.dao.VerifiedInstitutionalAffiliationDao;
-import org.pmiops.workbench.db.model.DbInstitution;
 import org.pmiops.workbench.db.model.DbUser;
-import org.pmiops.workbench.db.model.DbUserTermsOfService;
-import org.pmiops.workbench.db.model.DbVerifiedInstitutionalAffiliation;
-import org.pmiops.workbench.exceptions.BadRequestException;
-import org.pmiops.workbench.exceptions.NotFoundException;
-import org.pmiops.workbench.institution.InstitutionService;
-import org.pmiops.workbench.institution.VerifiedInstitutionalAffiliationMapper;
-import org.pmiops.workbench.model.InstitutionalRole;
 import org.pmiops.workbench.model.Profile;
-import org.pmiops.workbench.model.VerifiedInstitutionalAffiliation;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-@Service
-public class ProfileService {
-  private final FreeTierBillingService freeTierBillingService;
-  private final InstitutionDao institutionDao;
-  private final InstitutionService institutionService;
-  private final ProfileMapper profileMapper;
-  private final UserDao userDao;
-  private final UserService userService;
-  private final UserTermsOfServiceDao userTermsOfServiceDao;
-  private final VerifiedInstitutionalAffiliationDao verifiedInstitutionalAffiliationDao;
-  private final VerifiedInstitutionalAffiliationMapper verifiedInstitutionalAffiliationMapper;
+public interface ProfileService {
 
-  @Autowired
-  public ProfileService(
-      FreeTierBillingService freeTierBillingService,
-      InstitutionDao institutionDao,
-      InstitutionService institutionService,
-      ProfileMapper profileMapper,
-      UserDao userDao,
-      UserService userService,
-      UserTermsOfServiceDao userTermsOfServiceDao,
-      VerifiedInstitutionalAffiliationDao verifiedInstitutionalAffiliationDao,
-      VerifiedInstitutionalAffiliationMapper verifiedInstitutionalAffiliationMapper) {
-    this.freeTierBillingService = freeTierBillingService;
-    this.institutionDao = institutionDao;
-    this.institutionService = institutionService;
-    this.profileMapper = profileMapper;
-    this.userDao = userDao;
-    this.userService = userService;
-    this.userTermsOfServiceDao = userTermsOfServiceDao;
-    this.verifiedInstitutionalAffiliationDao = verifiedInstitutionalAffiliationDao;
-    this.verifiedInstitutionalAffiliationMapper = verifiedInstitutionalAffiliationMapper;
-  }
+  Profile getProfile(DbUser user);
 
-  public Profile getProfile(DbUser user) {
-    // Fetch the user's authorities, since they aren't loaded during normal request interception.
-    DbUser userWithAuthoritiesAndPageVisits =
-        userDao.findUserWithAuthoritiesAndPageVisits(user.getUserId());
-    if (userWithAuthoritiesAndPageVisits != null) {
-      // If the user is already written to the database, use it and whatever authorities and page
-      // visits are there.
-      user = userWithAuthoritiesAndPageVisits;
-    }
-
-    Profile profile = profileMapper.dbUserToProfile(user);
-
-    profile.setFreeTierUsage(freeTierBillingService.getCachedFreeTierUsage(user));
-    profile.setFreeTierDollarQuota(freeTierBillingService.getUserFreeTierDollarLimit(user));
-
-    verifiedInstitutionalAffiliationDao
-        .findFirstByUser(user)
-        .ifPresent(
-            verifiedInstitutionalAffiliation ->
-                profile.setVerifiedInstitutionalAffiliation(
-                    verifiedInstitutionalAffiliationMapper.dbToModel(
-                        verifiedInstitutionalAffiliation)));
-
-    Optional<DbUserTermsOfService> latestTermsOfServiceMaybe =
-        userTermsOfServiceDao.findFirstByUserIdOrderByTosVersionDesc(user.getUserId());
-    if (latestTermsOfServiceMaybe.isPresent()) {
-      profile.setLatestTermsOfServiceVersion(latestTermsOfServiceMaybe.get().getTosVersion());
-      profile.setLatestTermsOfServiceTime(
-          latestTermsOfServiceMaybe.get().getAgreementTime().getTime());
-    }
-
-    return profile;
-  }
-
-  public void validateInstitutionalAffiliation(Profile profile) {
-    VerifiedInstitutionalAffiliation verifiedInstitutionalAffiliation =
-        profile.getVerifiedInstitutionalAffiliation();
-
-    if (verifiedInstitutionalAffiliation == null) {
-      throw new BadRequestException("Institutional affiliation cannot be empty");
-    }
-
-    Optional<DbInstitution> institution =
-        institutionDao.findOneByShortName(
-            verifiedInstitutionalAffiliation.getInstitutionShortName());
-    if (!institution.isPresent()) {
-      throw new NotFoundException(
-          String.format(
-              "Could not find institution %s in database",
-              verifiedInstitutionalAffiliation.getInstitutionShortName()));
-    }
-    if (verifiedInstitutionalAffiliation.getInstitutionalRoleEnum() == null) {
-      throw new BadRequestException("Institutional role cannot be empty");
-    }
-    if (verifiedInstitutionalAffiliation.getInstitutionalRoleEnum().equals(InstitutionalRole.OTHER)
-        && (verifiedInstitutionalAffiliation.getInstitutionalRoleOtherText() == null
-            || verifiedInstitutionalAffiliation.getInstitutionalRoleOtherText().equals(""))) {
-      throw new BadRequestException(
-          "Institutional role description cannot be empty when institutional role is set to Other");
-    }
-
-    String contactEmail = profile.getContactEmail();
-    DbVerifiedInstitutionalAffiliation dbVerifiedAffiliation =
-        verifiedInstitutionalAffiliationMapper.modelToDbWithoutUser(
-            profile.getVerifiedInstitutionalAffiliation(), institutionService);
-    if (!institutionService.validateAffiliation(dbVerifiedAffiliation, contactEmail)) {
-      final String msg =
-          Optional.ofNullable(dbVerifiedAffiliation)
-              .map(
-                  affiliation ->
-                      String.format(
-                          "Contact email %s is not a valid member of institution '%s'",
-                          contactEmail, affiliation.getInstitution().getShortName()))
-              .orElse(
-                  String.format(
-                      "Contact email %s does not have a valid institutional affiliation",
-                      contactEmail));
-      throw new BadRequestException(msg);
-    }
-  }
+  void validateInstitutionalAffiliation(Profile profile);
 }
